puting openstreetmap function in an other file
